requirements.txt
numpy
pandas
datatable
matplotlib
pylab
opencv-python
Pillow
scikit-image
scipy
numba
torch
torchvision
tensorboardX
tensorflow>=2.0
keras
tqdm
datatable
scikit-image
scipy
numba
matplotlib
opencv-python
Pillow
torch
torchvision
tensorboardX
tensorflow>=2.0
keras
pylab
cv2
json
random
glob2